Pulled logic out of OrdersController
