+++ src/Foundation/Foundation.csproj
@@ -871,6 +871,7 @@
     <Compile Include="Features\MyAccount\ProfilePage\ProfilePageControllerService.cs" />
     <Compile Include="Features\MyAccount\SubscriptionDetail\SubscriptionDetailControllerService.cs" />
     <Compile Include="Features\MyOrganization\Budgeting\BudgetingControllerService.cs" />
+    <Compile Include="Features\MyOrganization\Orders\OrdersControllerService.cs" />
     <Compile Include="Infrastructure\PostClickInterceptor.cs" />
     <Compile Include="Features\CatalogContent\CatalogContentControllerBase.cs" />
     <Compile Include="Features\CatalogContent\Variation\VariationController.cs" />
